ConectaCareHC/requirements.txt
Flask
gunicorn
oracledb
requests
python-dotenv